lstsize return int-> size_t

+++ include/libft.h
@@ -6,7 +6,7 @@
 /*   By: bohlee <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#:+#+#+   +#+          */
 /*   Created: 2023/03/21 12:32:55 by bohlee            #+#    #+#             */
-/*   Updated: 2023/03/24 17:02:53 by bohlee           ###   ########.fr       */
+/*   Updated: 2024/04/22 11:34:24 by bohlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ void	ft_putendl_fd(char *s, int fd);
 void	ft_putnbr_fd(int n, int fd);
 t_list	*ft_lstnew(void *content);
 void	ft_lstadd_front(t_list **list, t_list *new);
-int		ft_lstsize(t_list *lst);
+size_t	ft_lstsize(t_list *lst);
 t_list	*ft_lstlast(t_list *lst);
 void	ft_lstadd_back(t_list **lst, t_list *new);
 void	ft_lstdelone(t_list *lst, void (*del)(void *));
